requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1

# AI / Agents
langchain==0.2.1
langchain-openai==0.1.8
langchain-community==0.2.1
langgraph==0.1.1
openai==1.30.1

# Data
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0

# Visualization
plotly==5.22.0
dash==2.17.1

# Vector Store / Memory
faiss-cpu==1.8.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
